// shared/electron-api-schema.ts
const updateStatePayload = z.object({
  status: z.enum(['idle', 'checking', 'available', 'downloading', 'downloaded', 'not-available']),
  currentVersion: z.string(),
  version: z.string().optional(),
  releaseNotes: z.string().optional(),
  percent: z.number().optional(),
  message: z.string().optional(),
})
export type UpdateStatePayload = z.infer<typeof updateStatePayload>